Update README.md
Adding the setup instructions from my personal story mode patch fork so that more new devs can maybe attempt to assist with the disassembly further. This README may change down the road to be more in-line with the main PRET repo to see if they possibly allow me to merge the Build Instructions section to their README to help the project continue to grow.