bank03.asm Func/Data/Text Names
Adding a hopefully relatively consistent naming scheme for some more func/data/text that was not previously named. Trying to keep consistent with final Gold/Yellow/Red  naming schemes as much as possible. This should hopefully be the final bits of what was necessary for the pull request to the PRET repo.