bank03.asm Function Descriptions & Translations (#110)
* bank03.asm Function Descriptions & Translations

Comments added to many functions and text strings that were not fully disassembled. Also translated to some of the text within the file related to the functions that have been researched.

* Update bank03.asm

Beginning to rename functions/text names to match gold/yellow/red depending on which final game was closest to the code/text in the demo. Also updating/adding some translations.

* Update bank03.asm

Continuing to rename functions/text names to match gold/yellow/red depending on which final game was closest to the code/text in the demo. Also updated/added some more translations.

* Update bank03.asm - Missed One

Missed this change in the last commit since I apparently didn't save the file. Whoops. Last commit for this session.

* Update README.md

Adding the setup instructions from my personal story mode patch fork so that more new devs can maybe attempt to assist with the disassembly further. This README may change down the road to be more in-line with the main PRET repo to see if they possibly allow me to merge the Build Instructions section to their README to help the project continue to grow.

* bank03.asm Func/Data/Text Names

Adding a hopefully relatively consistent naming scheme for some more func/data/text that was not previously named. Trying to keep consistent with final Gold/Yellow/Red  naming schemes as much as possible. This should hopefully be the final bits of what was necessary for the pull request to the PRET repo.

* bank03.asm 2 Reverts, Translations, & Comments

I reverted 2 text name changes as they might be incorrect assumptions based on recent findings while continuing to translate text within the file. I also added a comment or two on some functions that I need to research further in the future when I have more time.

* Revert "Update README.md"

This reverts commit 4e9baa927b21d8db44bfad4dd3a9372731b5f70c.

* Revert a few function names

Reverting a few function names that I discovered I named poorly. Will be trying to make more changes based on the feedback received by the team as soon as I can. Just want to get these few reverted for now so that I can properly research them further without reverting all of the other changes I did in a previous commit that were actually helpful.

* Corrections & Additions from vulcandth

Making corrections and additional name changes based on the amazing suggestions from vulcandth.

* Vulcandth bank03.asm sugggestions & build fix

This should fix the build error as well as make the suggested corrections from Vulcandth so that it should be compatible. I hope. Lol.